backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
Pillow>=10.0.0
numpy>=1.24.0
pymongo>=4.6.0
python-dotenv>=1.0.0
reportlab>=4.0.0
aiofiles>=24.0.0
